# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QCS Cargo Core API
  version: 1.0.0
  summary: Core API surface for auth, customer operations, tracking, sessions, and admin monitoring.
  description: |
    Practical OpenAPI spec for MISS-037 based on current handler behavior in `internal/api`.
    All endpoints are under `/api/v1`.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.qcs-cargo.com
    description: Production (example)
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register user account
      operationId: postAuthRegister
      security: []
      description: Creates a customer account and triggers email verification flow.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRegisterResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email is already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailExists:
                  value:
                    error:
                      code: EMAIL_EXISTS
                      message: email already registered
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/magic-link/request:
    post:
      tags: [Auth]
      summary: Request magic-link sign-in email
      operationId: postAuthMagicLinkRequest
      security: []
      description: |
        Login-equivalent step 1. Returns the same success message even when the email does not exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequestRequest'
      responses:
        '200':
          description: Request accepted (enumeration-safe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/magic-link/verify:
    post:
      tags: [Auth]
      summary: Verify magic-link token and create session
      operationId: postAuthMagicLinkVerify
      security: []
      description: |
        Login-equivalent step 2. Returns access token and sets refresh cookie (`qcs_refresh`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkVerifyRequest'
      responses:
        '200':
          description: Signed in
          headers:
            Set-Cookie:
              description: HttpOnly refresh token cookie `qcs_refresh`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid or expired magic link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidLink:
                  value:
                    error:
                      code: INVALID_LINK
                      message: invalid or expired magic link
        '403':
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailNotVerified:
                  value:
                    error:
                      code: EMAIL_NOT_VERIFIED
                      message: Please verify your email before signing in.
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      operationId: postAuthLogout
      description: |
        Revokes refresh session (if cookie exists), best-effort blacklists access token (if bearer token provided),
        and clears `qcs_refresh` cookie.
      security:
        - bearerAuth: []
        - refreshCookie: []
      responses:
        '204':
          description: Logged out
          headers:
            Set-Cookie:
              description: Clears `qcs_refresh`
              schema:
                type: string
        '403':
          description: CSRF check failed when refresh cookie is present and origin/referer is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                csrfFailed:
                  value:
                    error:
                      code: CSRF_FAILED
                      message: Origin not allowed
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email with token
      operationId: postAuthVerifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidToken:
                  value:
                    error:
                      code: INVALID_TOKEN
                      message: token invalid or expired
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      operationId: postAuthResendVerification
      security: []
      description: Always returns a generic success message for unknown/already-verified emails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '200':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token in `Authorization: Bearer <token>`."
    refreshCookie:
      type: apiKey
      in: cookie
      name: qcs_refresh
      description: |
        HttpOnly refresh-session cookie used by auth refresh/logout flows.
        For state-changing requests with this cookie, middleware enforces allowed Origin/Referer (CSRF protection).
  responses:
    ValidationError:
      description: Request body/params failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid body
...
    RateLimited:
      description: Too many auth attempts or temporary auth lockout
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            limited:
              value:
                error:
                  code: RATE_LIMITED
                  message: Too many requests from this IP. Please try again later.
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: An unexpected error occurred
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              description: Optional structured details
    MessageResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [message]
          properties:
            message:
              type: string
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
        - suite_code
        - storage_plan
        - free_storage_days
        - email_verified
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
          enum: [customer, staff, admin]
        avatar_url:
          type: string
        suite_code:
          type: string
        address_street:
          type: string
        address_city:
          type: string
        address_state:
          type: string
        address_zip:
          type: string
        storage_plan:
          type: string
        free_storage_days:
          type: integer
        email_verified:
          type: boolean
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthRegisterRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          description: Optional. If provided, must satisfy password policy.
    AuthRegisterResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [user, message]
          properties:
            user:
              $ref: '#/components/schemas/User'
            message:
              type: string
    MagicLinkRequestRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          description: Ignored by backend today.
        redirectTo:
          type: string
          description: Optional post-login redirect path.
    MagicLinkVerifyRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    ResendVerificationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AuthSessionResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [user, access_token, session_id, expires_in]
          properties:
            user:
              $ref: '#/components/schemas/User'
            access_token:
              type: string
            session_id:
              type: string
              description: Current refresh-session identifier used by session-management UI.
            expires_in:
              type: integer
              description: Access-token lifetime in seconds.
...